SIGPIPE & SIGSEGV + Bug Fix
Due to network issue my server was getting SIGPIPE and this was causing a crash
Blocked SIGPIPE

Another issue was that the server was crashing with SIGSEGV.
I have written a controlled stack-unrolling patch to ensure our server
doesn't crash. Instead it just resets to start point.

Also there was in issue with get-list in items as 5MB buffer was taken on stack
as static. Allocation failed and caused re-curring crash. Moved that to heap.

Signed-off-by: khubaibumer <khubaib.umer@aircod.com>